simplified reducers and store
