+++ client/store/index.js
@@ -3,17 +3,13 @@ import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
-import allProducts from './all-products'
-import singleProduct from './single-product'
+import products from './products'
 import cart from './cart'
-import orderHistory from './order-history'
 
 const reducer = combineReducers({
   user,
-  allProducts,
-  singleProduct,
-  cart,
-  orderHistory
+  products,
+  cart
 })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
@@ -22,7 +18,5 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
-export * from './all-products'
-export * from './single-product'
+export * from './products'
 export * from './cart'
-export * from './order-history'
