+++ package.json
@@ -61,6 +61,7 @@
     "react-redux": "^7.0.1",
     "react-router": "^5.0.1",
     "react-router-dom": "^5.0.0",
+    "react-stripe-checkout": "^2.6.3",
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
