+++ server/index.js
@@ -11,6 +11,7 @@ const PORT = process.env.PORT || 8080
 const app = express()
 const socketio = require('socket.io')
 module.exports = app
+//const stripe = require('stripe')('sk_test_4eC39HqLyjWDarjtT1zdp7dc')
 
 // This is a global Mocha hook, used for resource cleanup.
 // Otherwise, Mocha v4+ never quits after tests.
